adicionado paamentos e schemas de arquivos produtos !

+++ schemas/pedido_schema.py
@@ -1,10 +1,11 @@
+from decimal import Decimal
 from pydantic import BaseModel
 
 class PedidoBase(BaseModel):
     usuario_id: int
     unidade_id: int
     data_pedido: str
-    total: float
+    total: Decimal
     status = enumerate
     canal_pedido: enumerate
     
@@ -12,5 +13,5 @@ class ItemPedidoBase(BaseModel):
     pedido_id: int
     produto_id: int
     quantidade: int
-    preco_unitario: float
+    preco_unitario: Decimal
     subtotal: float
